Fix detection of Asterisk version 11 and up.
As it stands, Asterisk 11 removed version information from its header
file(s), only leaving two functions ast_get_version() and
ast_get_version_num().

Those would generally suffice/be an equivalent replacement of the old
literal string macros, but there's no way of using the functions due to
a missing implementation on Asterisk's side.

Within Asterisk, both functions are implemented in an autogenerated(!)
file named version.c, which can only be used at Asterisk build time.

As chan_capi is an external module/channel and the functions are not
exposed in a library, there's no way of using
ast_get_version()/ast_get_version_num().

Instead, the officially sanctioned way of determining the current
installed Asterisk version is by running ``asterisk -V'' and parsing the
output (which may be unstable, so we will have to keep an eye on that.)